refactor minitouch, add a super class and a new touch type for compatibility with android10
